Move `mill.BuildInfo` to`mill.main.BuildInfo` (#2537)
We currently encourage `import mill._`, which causes problems when
people want to also `import mill.contrib.buildinfo.BuildInfo`. There's a
workaround to `import mill.{BuildInfo => MillBuildInfo, _}`, which we do
in several places in our own codebase, but it's a pain in the neck and
an unnecessary hurdle for users to cross. The majority of users would
not need to access `mill.BuildInfo` at all, so putting it in the default
namespace just adds clutter.

The `mill.BuildInfo` is used throughout `mill.scalalib` and
`mill.scalajslib` and so on, but is not used in `main/`, so I put it in
as downstream a module as possible which turns out to be
`mill.main.BuildInfo`
