+++ scalalib/src/mill/scalalib/Lib.scala
@@ -4,6 +4,7 @@ package scalalib
 import coursier.util.Task
 import coursier.{Dependency, LocalRepositories, Repositories, Repository, Resolution}
 import mill.api.{Ctx, Loose, PathRef, Result}
+import mill.main.BuildInfo
 import mill.modules.Util
 import mill.scalalib.api.ZincWorkerUtil
