+++ scalalib/src/mill/scalalib/TestModule.scala
@@ -171,7 +171,7 @@ object TestModule {
     override def testFramework: T[String] = "mill.testng.TestNGFramework"
     override def ivyDeps: T[Agg[Dep]] = T {
       super.ivyDeps() ++ Agg(
-        ivy"com.lihaoyi:mill-contrib-testng:${mill.BuildInfo.millVersion}"
+        ivy"com.lihaoyi:mill-contrib-testng:${mill.api.BuildInfo.millVersion}"
       )
     }
   }
